==> 02-Book-review-and-ratings.ipynb <==
# Book review and ratings by user

import goodreads_api_client as gr
from goodreads import client
import pickle
import requests
from bs4 import BeautifulSoup
import xmltodict
import pprint
import json
import os
import re
import pandas as pd
import numpy as np
#!pip install selenium
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
chromedriver = "/Applications/chromedriver" # path to the chromedriver executable
os.environ["webdriver.chrome.driver"] = chromedriver

**Goodreads API credentials and wrappers**

api_key = 'Jjd519OAM82QfNQGepZsw'
api_secret = 'JY9m0mBvG7U1Ws4AUkpMucbwCNi6tndGTIqdASYo'

client = gr.Client(developer_key=api_key)

from goodreads import client
gc = client.GoodreadsClient(api_key, api_secret)

**Define functions to access APIs and scrape**

def get_info(url):
    key = api_key
    response = requests.get(url, params={'key': api_key})
    #soup = BeautifulSoup(response.content, "lxml-xml")
    results = xmltodict.parse(response.content)
    #results = json.dumps(results)
    return(results['GoodreadsResponse'])

def get_soup(url):
    response = requests.get(url)
    page = response.text
    soup = BeautifulSoup(page, "lxml")
    #print(soup.prettify())
    return(soup)

**List of book id's on fantasy genre lists**

with open("book_ids.pkl", "rb") as f:
    fantasy_ids = pickle.load(f)
fantasy_ids = list(set(fantasy_ids))

**Define a function to get data for top reviews of a book by id**

def get_book_review_dict_by_user(i):
    d = {}
#for i in fantasy_ids[0:1]:
    d[i] = {}

    url = 'https://www.goodreads.com/book/show/'+i+'.xml'

    soup = get_soup(url)
    review_data = soup.find_all("div", {"class" :"review"})

    for r in review_data:
        try:
            userid = r.find("a", {"class" :"user"})['href'].split('show/')[1].split('-')[0]
            d[i][userid] = {}
            try:
                d[i][userid]['user'] = r.find("a", {"class" :"user"}).text
            except:
                pass
            try:
                d[i][userid]['review'] = r.find("div", {"class" :"reviewText"}).text.strip()
            except:
                pass
            try:
                d[i][userid]['rating'] = r.find("span", {"class" :"staticStars"}).text
            except:
                pass
                #d[i]['user_page'] = r.find("a", {"class" :"user"})['href']
            try:
                d[i][userid]['review_date'] = r.find("a", {"class" :"reviewDate"}).text            
            except:
                pass
        except:
            pass  
    return(d)

**Iterate through list of book id's. Stop every 1000 items to pickle the results. Save the final iteration variable as well**

len(fantasy_ids)

**Reviews for first 10000 books**

for count, bookid in enumerate(fantasy_ids[0:10001]): 
    if count == 0:
        print(bookid)
    else:
        if count % 1000 == 0:
            print(count, bookid)

book_review_dict = get_book_review_dict_by_user(fantasy_ids[0])

iters = 0
for count, bookid in enumerate(fantasy_ids[0:10001]): 
    if count == 0:
        pass
    else:
        book_review_dict.update(get_book_review_dict_by_user(bookid))
    # Every 1000 books stop and pickle the dictioary
        if count % 1000 == 0:
            with open('book_reviews_by_user.pkl', 'wb') as f:
                pickle.dump(book_review_dict, f, pickle.HIGHEST_PROTOCOL)
            print(count)
            iters = count
            with open('book_review_iter_status.pkl', 'wb') as f:
                pickle.dump(iters, f)


with open('book_reviews_by_user.pkl', 'wb') as f:
    pickle.dump(book_review_dict, f, pickle.HIGHEST_PROTOCOL)

test = pd.read_pickle('book_reviews_by_user.pkl')

len(book_review_dict)

**Remaining 30k+ reviews**

#book_review_dict = pd.read_pickle('book_reviews_by_user.pkl')

iters = len(book_review_dict)

for count, bookid in enumerate(fantasy_ids[iters:]): 
    if count == 0:
        pass
    else:
        book_review_dict.update(get_book_review_dict_by_user(bookid))
    # Every 1000 books stop and pickle the dictioary
        if count % 1000 == 0:
            with open('book_reviews_by_user.pkl', 'wb') as f:
                pickle.dump(book_review_dict, f, pickle.HIGHEST_PROTOCOL)
            print(count+15000)
            iters = count
            with open('book_review_iter_status.pkl', 'wb') as f:
                pickle.dump(iters, f)


len(book_review_dict)

list(book_review_dict.keys())[-1], fantasy_ids[26001]

**Rearrange dictionary structure for mondgodb**

# Turn dictionary into a list
book_reviews_10k = []
for k, v in book_review_dict.items():
    book_reviews_10k.append({'bookid': k, 'reviews': v})

# Move userid to inside the review data rather than key to review data
reviews = []
for i, b in enumerate(book_reviews_10k):

    for userid in book_reviews_10k[i]['reviews'].keys():
        
        book_reviews_10k[i]['reviews'][userid]['userid'] = userid
        
# Final list of dictionaries
book_reviews = []

fix_iters = 0

for i, book in enumerate(book_reviews_10k):
    reviews = []
    for review in book['reviews'].values():
        reviews.append(review)
    book_reviews.append({'bookid': book['bookid'], 'reviews': reviews})
        
    # Every 1000 books stop and pickle
    if i % 1000 == 0:
        with open('book_reviews_dict_first10k.pkl', 'wb') as f:
            pickle.dump(book_reviews, f, pickle.HIGHEST_PROTOCOL)
        iters = count
        with open('book_review_fix_iter_status.pkl', 'wb') as f:
            pickle.dump(fix_iters, f)


**END NOTEBOOK**

***
