==> 11 - EDA.ipynb <==
# EDA

from bokeh.plotting import figure
from bokeh.embed import components
from bokeh.models import ColumnDataSource, NumeralTickFormatter
from bokeh.palettes import Spectral6, Pastel1
from bokeh.models import ColumnDataSource
from bokeh.plotting import show, output_notebook, figure as bf
from bokeh.plotting import figure
from bokeh.models.glyphs import  Text
from bokeh.layouts import row, widgetbox
from bokeh.models import CustomJS, Slider
from bokeh.io import curdoc
from bokeh.layouts import row, column
from bokeh.models.widgets import Slider, TextInput
from bokeh.models import HoverTool
from bokeh.layouts import gridplot
from bokeh.models import ColumnDataSource, ranges, LabelSet

import math

import pandas as pd

colors = ['blue','green','red']
source = ColumnDataSource(data=dict(x=[0.2, 0.1, 0.7], y=['YA','epic','vamp'], colors=colors))
    
plot = figure(x_range = (0,1), y_range = source.data['y'], title="Destinations outside of USA/Canada", 
                    plot_height=200,plot_width=500)

plot.hbar(y = 'y', left=0, right='x', height=0.9, color='colors', source=source)
    #plot.add_tools(HoverTool(tooltips=[("LOCATION", "@x"), ("PCT BOOKED", "@counts %")]))
    #plot.xaxis.major_label_orientation = math.pi/4
    #plot.xaxis.major_label_text_font_size = "10pt"
    #plot.yaxis.minor_tick_line_color = None
    #plot.xaxis.axis_label = 'Destination country'
    #plot.xaxis.axis_label_text_font_size = "20pt"
plot.yaxis.axis_label = 'Sub-genre'
show(plot)

books = pd.read_pickle('final_book_topics.pkl')

print(list(books['title'][books['bookid']=='20518872']))

books.head()

print(str(books['title'][books['bookid']=='1']))

books.iloc[7407,3]

**Cosine similarity**

vec1 = np.array([[1,1,0,1,1]])
vec2 = np.array([[0,1,0,1,1]])
#print(cosine_similarity([vec1, vec2]))
#print(cosine_similarity(vec1, vec2))
vec1

title='Harry Potter and the Half-Blood Prince (Harry Potter, #6)'
books[books['title']==title].index[0]

from sklearn.metrics.pairwise import cosine_similarity
import numpy as np


vec1=np.array([books[['topic_1','topic_2','topic_3','topic_4','topic_5']].iloc[1,:]])
vec2=np.array([books[['topic_1','topic_2','topic_3','topic_4','topic_5']].iloc[2,:]])
print(cosine_similarity(vec1, vec2))


matrix = books[['topic_1','topic_2','topic_3','topic_4','topic_5']].values
matrix[0]

from sklearn.metrics.pairwise import cosine_similarity
cosine_similarity(matrix[0:1], matrix)


a = np.array([9, 4, 4, 3, 3, 9, 0, 4, 6, 0])
a.argsort()[-2:]

import string
import re
matrix = books[['topic_1','topic_2','topic_3','topic_4','topic_5']].values
def get_recs(title):
    i = books[books['title']==title].index[0] # i is index corresponding to that title
    
    cosines = cosine_similarity(matrix[i:i+1], matrix)
    cosines = cosines.tolist()[0]
    cosines = [float(X) for X in cosines]
    sim_book_indexes = sorted(range(len(cosines)), key=lambda i: cosines[i])[-20:]
    #cosines = sorted(cosines, key=lambda x: x[1], reverse=True)
    titles = []
    for idx in sim_book_indexes:
        if books['title'].iloc[idx] == title:
            pass
        else:
            titles.append(books['title'].iloc[idx])
    for t in titles:
        for n in range(2,20):
            if '#'+str(n) in t:
                titles.remove(t)
    #return (" \n".join(map(str,titles)))
    return(titles)

get_recs('Harry Potter and the Half-Blood Prince (Harry Potter, #6)')

print(books['title'].iloc[7407])

books['bookid'][7407]

i = books[books['title']==title].index[0]
i

sorted(range(len(cos)), key=lambda i: cos[i])[-15:]

cos[7407]

cos.index(max(cos))

HP = sorted(HP, key=lambda x: x[1], reverse=True)
for book in HP[0:25]:
    #print(book)
    #print(book[0])
    print(books['title'].iloc[book[0]])

userid= 10015197
df = pd.read_pickle('flask/user_product_long.pkl')
book_topics = pd.read_pickle('final_book_topics.pkl')
topic_dict = {'topic_1':'young adult',
              'topic_2':'epic fantasy/fairytale',
              'topic_3':'supernatural',
              'topic_4':'sci-fi',
              'topic_5':'urban fantasy'
             }
df = pd.merge(df, book_topics[['bookid','topic_1','topic_2','topic_3','topic_4','topic_5','topic','topic_name']], how='left', on='bookid')
books_read = df[['rating','title','bookid','topic_1','topic_2','topic_3','topic_4','topic_5']][df['userid']==int(userid)].drop_duplicates()



import pickle
with open("flask/recommender_algorithm.pkl", "rb") as f:
    algo = pickle.load(f)
with open("flask/book_titles_map.pkl", 'rb') as f4:
    fantasy_books = pickle.load(f4)
book_rating_aggs = pd.read_pickle('flask/fantasy_books_info_dataframe.pkl')
books_read = pd.merge(books_read, book_rating_aggs[['bookid','average_rating']], on = 'bookid',how='left')
books_not_read = list(set(df['bookid'].unique()) - set(list(books_read)))
estimates = []
for bookid in books_not_read:
    uid = int(userid)
    iid = bookid 
    estimates.append(algo.predict(uid, iid, verbose=False))

book_recs = pd.DataFrame(estimates).sort_values(by='est', ascending = False)
book_recs['title'] = book_recs['iid'].map(fantasy_books)
book_recs['est'] = [round(X, 2) for X in book_recs['est']]

book_recs = book_recs[(book_recs['est'] == 5.0)]
for i in range(2,20):
    book_recs = book_recs[book_recs['title'].str.contains('#'+str(i))==False]
book_recs = pd.merge(book_recs, 
                     book_rating_aggs[['bookid','average_rating','ratings_count']],
                     left_on = 'iid',
                     right_on = 'bookid',
                     how = 'left'
                    )

book_recs.head()

book_recs = pd.merge(book_recs, book_topics[['bookid','topic_1','topic_2','topic_3','topic_4','topic_5','topic','topic_name']], 
                        left_on = 'iid',
                        right_on = 'bookid',
                        how = 'left'
                        )

data = pd.DataFrame(books_read[['topic_1','topic_2',
                                'topic_3','topic_4','topic_5']].mean()).reset_index()
data.max()['index']

favorite_topic = books_read.group_by

from surprise import KNNBasic
from surprise import Dataset
from surprise import Reader
df = pd.read_pickle('flask/user_product_long.pkl')

reader = Reader(rating_scale=(1, 5))

# The columns must correspond to user id, item id and ratings (in that order).
data = Dataset.load_from_df(df[['userid', 'bookid', 'rating']], reader)

sim_options = {
    'name': 'cosine',
    'user_based': False
}
 
knn = KNNBasic(sim_options=sim_options)
predictions = knn.test(data)

**Topic EDA**

for col in ['topic_1','topic_2','topic_3','topic_4','topic_5']:
    books[col] = [X[1]  if X is not None else 0 for X in books[col]]

pd.to_pickle(books,'final_book_topics.pkl')

topic_dict = {'topic_1':'young adult',
              'topic_2':'epic /fairytale',
              'topic_3':'supernatural',
              'topic_4':'sci-fi',
              'topic_5':'urban fantasy'
             }

data = pd.DataFrame(books[['topic_1','topic_2',
       'topic_3','topic_4',
       'topic_5']].mean()).reset_index()
data.columns=['topic','values']
data['topic'] = data['topic'].map(topic_dict)
data['values'] = [X*100 for X in data['values']]
data = data.sort_values(by='values', ascending = False)
data

colors = ['orangered','tomato','coral','darkorange','orange']
source = ColumnDataSource(data=dict(x=data['values'], y=data['topic'], colors=colors))


plot = figure(y_range = (0,data['values'].max()+5), x_range = source.data['y'], 
              plot_height=400,plot_width=700)

plot.vbar(x = 'y', bottom=0, top='x', width=0.7, color='colors', source=source)

plot.add_tools(HoverTool(tooltips=[("topic", "@y"), ("Percent", "@x{int}%")]))

plot.xaxis.major_label_orientation = math.pi/4
plot.xaxis.major_label_text_font_size = "14pt"
plot.xaxis.major_label_text_font_style = "bold"
plot.xaxis.axis_label_text_font_size = "16pt"
plot.xaxis.axis_label_text_font_style = "normal"

plot.yaxis.axis_label = 'Percent of books'
plot.yaxis.minor_tick_line_color = None
plot.yaxis.major_tick_line_color = None
plot.yaxis.major_label_text_font_size = "10pt"

plot.yaxis.axis_label_text_font_size = "16pt"
plot.yaxis.axis_label_text_font_style = "bold"
plot.xaxis.axis_label_text_color = "white"
plot.yaxis.axis_label_text_color = "white"
plot.yaxis.major_label_text_color = "white"
plot.xaxis.major_label_text_color = "white"

plot.xgrid.grid_line_color = None
plot.ygrid.grid_line_color = None
plot.xaxis.axis_line_color = "white"
plot.yaxis.axis_line_color = "white"

plot.background_fill_alpha = 0
plot.border_fill_alpha = 0

show(plot)

source.data

list(book_topics.columns)

df = pd.read_pickle('flask/user_product_long.pkl')
book_topics = pd.read_pickle('final_book_topics.pkl')
df = pd.merge(df, book_topics, how='left', on='bookid')
df.head()

df = pd.read_pickle('flask/user_product_long.pkl')

book_topics = pd.read_pickle('flask/final_book_topics.pkl')
topic_dict = {'topic_1':'young adult',
              'topic_2':'epic fantasy/fairytale',
              'topic_3':'supernatural',
              'topic_4':'sci-fi',
              'topic_5':'urban fantasy'
             }
df = pd.merge(df, book_topics[['bookid','topic_1','topic_2','topic_3','topic_4','topic_5','topic','topic_name']], how='left', on='bookid')


matrix = df[['topic_1','topic_2','topic_3','topic_4','topic_5']].values


def get_books(title):
    i = df[df['title']==title].index[0] # i is index corresponding to that title
        
    cosines = cosine_similarity(matrix[i:i+1], matrix)
    cosines = cosines.tolist()[0]
    cosines = [float(X) for X in cosines]
    sim_book_indexes = sorted(range(len(cosines)), key=lambda i: cosines[i])[-20:]
        #cosines = sorted(cosines, key=lambda x: x[1], reverse=True)
    titles = []
    for idx in sim_book_indexes:
        if books['title'].iloc[idx] == title:
            pass
        else:
            titles.append(books['title'].iloc[idx])
    for t in titles:
        for n in range(2,20):
            if '#'+str(n) in t:
                titles.remove(t)
        #return (" \n".join(map(str,titles)))
    return(titles)

results = get_books(title)

title = "Fight Club"
i = df[df['title']==title].index[0]

df.iloc[1463]

book_topics = pd.read_pickle('final_book_topics.pkl')

cols = ['topic_1','topic_2','topic_3','topic_4','topic_5']
matrix = book_topics[cols].values


print(len(book_topics),len(matrix))

title = "Matilda"
i = book_topics[book_topics['title']==title].index[0] # i is index corresponding to that title  
cosines = cosine_similarity(matrix[i:i+1], matrix)
cosines = cosines.tolist()[0]
sim_book_indexes = sorted(range(len(cosines)), key=lambda i: cosines[i])[-20:]
sim_book_indexes


book_topics.iloc[12563]

cols = ['topic_1','topic_2','topic_3','topic_4','topic_5']
matrix = book_topics[cols].values
def get_books(title):
    i = book_topics[book_topics['title']==title].index[0] # i is index corresponding to that title      
    cosines = cosine_similarity(matrix[i:i+1], matrix)
    cosines = cosines.tolist()[0]
    sim_book_indexes = sorted(range(len(cosines)), key=lambda i: cosines[i])[-20:]
        #cosines = sorted(cosines, key=lambda x: x[1], reverse=True)
    titles = []
    for idx in sim_book_indexes:
        if book_topics['title'].iloc[idx] == title:
            pass
        else:
            titles.append(book_topics['title'].iloc[idx])
    for t in titles:
        for n in range(2,20):
            if '#'+str(n) in t:
                titles.remove(t)
        #return (" \n".join(map(str,titles)))
    return(set(titles))

get_books('Matilda')

df = df.drop_duplicates()
df[df[['topic_1','topic_2','topic_3','topic_4','topic_5']].isnull().values == False]
